// frontend-app/src/apiClient/models/UserTokenHolder.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UserTokenHolder = {
	cognito_user_id: string;
	account_address: string;
	first_name: string;
	last_name: string;
	email: string;
	token_id: string;
	token_contract_address: string;
	balance_frozen: string;
	balance_unfrozen: string;
	balance_total: string;
};